FIX: small fixes for non-uniform scale

+++ UnityPackage/Runtime/Scripts/WaterSurface.cs
@@ -35,7 +35,7 @@ namespace WaterWaveSurface
             cam.transform.rotation = Quaternion.Euler(90,0, -transform.rotation.eulerAngles.y);
             cam.transform.localScale = Vector3.one;
             cam.orthographicSize = m_settings.environment.size.y;
-            cam.aspect = m_settings.environment.size.x / m_settings.environment.size.y;
+            cam.aspect = m_settings.environment.size.x / (float)m_settings.environment.size.y;
             cam.nearClipPlane = 0f;
             cam.farClipPlane = transform.lossyScale.y;
